requirements.txt
gunicorn
flask
flask-cors
boto3
psycopg2-binary
openai
pinecone
sentence-transformers
python-dotenv
nltk
paramiko
pandas
openpyxl
rank-bm25